public index / favicon cleanup all apps
