+++ apps/data-product-detail/public/index.html
@@ -4,7 +4,7 @@
     <title>NEON | Data Product</title>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="theme-color" content="#000000">
+    <meta name="theme-color" content="#002c77">
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico?201912">
     <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons"/>
