+++ apps/data-product-detail/public/manifest.json
@@ -10,6 +10,6 @@
   ],
   "start_url": "./index.html",
   "display": "standalone",
-  "theme_color": "#000000",
+  "theme_color": "#002c77",
   "background_color": "#ffffff"
 }
