+++ apps/explore-data-products/public/manifest.json
@@ -5,11 +5,11 @@
     {
       "src": "favicon.ico",
       "sizes": "16x16 32x32 48x48",
-      "type": "image/png"
+      "type": "image/x-icon"
     }
   ],
   "start_url": "./index.html",
   "display": "standalone",
-  "theme_color": "#000000",
+  "theme_color": "#002c77",
   "background_color": "#ffffff"
 }
