+++ apps/taxon/.env.development
@@ -23,4 +23,4 @@ REACT_APP_NEON_AUTH_USERINFO="/auth0/userinfo"
 # Override for development purposes to point at various environments
 # Otherwise will derive from where the app is running
 #REACT_APP_NEON_ALLOW_PROD_HOST_OVERRIDE="false"
-#REACT_APP_NEON_API_ROOT_PATH_OVERRIDE="https://cert-data.neonscience.org"
+REACT_APP_NEON_API_ROOT_PATH_OVERRIDE="https://cert-data.neonscience.org"
