+++ apps/taxon/public/index.html
@@ -1,33 +1,36 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
+    <title>NEON | Taxonomic Lists</title>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="theme-color" content="#000000">
+    <meta name="theme-color" content="#002c77">
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
-    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
-	<title>Taxonomic Lists</title>
-	<!-- Google Tag Manager -->
+    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico?v=201912">
+    <!-- jQuery needed for drupal header -->
+    <script src="https://code.jquery.com/jquery-3.5.0.min.js" integrity="sha256-xNzN2a4ltkB44Mc/Jz3pT4iU1cmeR0FkXs4pru/JxaQ=" crossorigin="anonymous"></script>
+    <!-- Google Tag Manager -->
+    <script>gtmDataLayer=[{'pageCategory':'Taxonomic Lists'}];</script>
     <script>(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
-		new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
-		j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
-		'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-		})(window,document,'script','dataLayer','GTM-K4S83R2');
-	</script>
-	<!-- End Google Tag Manager -->
-	<script>
-		window.NEON_SERVER_DATA = "__NEON_SERVER_DATA__";
-	</script>
+      new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
+      j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
+      'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
+      })(window,document,'script','gtmDataLayer','GTM-K4S83R2');
+    </script>
+    <!-- End Google Tag Manager -->
+    <script>
+      window.NEON_SERVER_DATA = "__NEON_SERVER_DATA__";
+    </script>
   </head>
   <body>
     <noscript>
       You need to enable JavaScript to run this app.
-	</noscript>
-	<!-- Google Tag Manager (noscript) -->
+    </noscript>
+    <!-- Google Tag Manager (noscript) -->
     <noscript><iframe src="https://www.googletagmanager.com/ns.html?id=GTM-K4S83R2"
 		height="0" width="0" style="display:none;visibility:hidden"></iframe>
-	</noscript>
-	<!-- End Google Tag Manager (noscript) -->
+    </noscript>
+    <!-- End Google Tag Manager (noscript) -->
     <div id="root"></div>
   </body>
 </html>
