+++ apps/taxon/public/manifest.json
@@ -1,15 +1,15 @@
 {
   "short_name": "Taxonomic Lists",
-  "name": "NEON Portal Taxonomic Lists",
+  "name": "NEON | Taxonomic Lists",
   "icons": [
     {
       "src": "favicon.ico",
-      "sizes": "64x64 32x32 24x24 16x16",
+      "sizes": "16x16 32x32 48x48",
       "type": "image/x-icon"
     }
   ],
   "start_url": "./index.html",
   "display": "standalone",
-  "theme_color": "#000000",
+  "theme_color": "#002c77",
   "background_color": "#ffffff"
 }
