+++ apps/taxon/src/components/app/App.jsx
@@ -1,0 +1,28 @@
+import React from "react";
+
+import InfoPresentation from "../presentations/InfoPresentation";
+import ControlPresentation from "../presentations/ControlPresentation";
+import ColumnManagerContainer from "../containers/ColumnManagerContainer";
+import DataTableContainer from "../containers/DataTableContainer";
+
+import NeonPage from "portal-core-components/lib/components/NeonPage";
+
+const App = () => {
+  const breadcrumbs = [
+    { name: "Taxonomic Lists" },
+  ];
+  return (
+    <NeonPage
+      title="Taxonomic Lists"
+      breadcrumbs={breadcrumbs}
+      useCoreHeader
+    >
+      <InfoPresentation />
+      <ControlPresentation />
+      <ColumnManagerContainer />
+      <DataTableContainer />
+    </NeonPage>
+  );
+};
+
+export default App;
