--- apps/taxon/src/components/app/AppLayout.css
@@ -1,4 +1,0 @@
-.taxon-container .paper-container {
-  padding: 15px;
-  background-color: #ffffff;
-}
