+++ apps/taxon/src/index.js
@@ -13,8 +13,6 @@ import ReactDOM from "react-dom";
 
 import Root from "./Root";
 
-import "./index.css";
-
 ReactDOM.render(
   <Root />,
   document.getElementById("root")
